Prettify all contracts under the same rules

+++ contracts/ethregistrar/StringUtils.sol
@@ -7,13 +7,13 @@ library StringUtils {
      * @param s The string to measure the length of
      * @return The length of the input string
      */
-    function strlen(string memory s) internal pure returns (uint) {
-        uint len;
-        uint i = 0;
-        uint bytelength = bytes(s).length;
-        for(len = 0; i < bytelength; len++) {
+    function strlen(string memory s) internal pure returns (uint256) {
+        uint256 len;
+        uint256 i = 0;
+        uint256 bytelength = bytes(s).length;
+        for (len = 0; i < bytelength; len++) {
             bytes1 b = bytes(s)[i];
-            if(b < 0x80) {
+            if (b < 0x80) {
                 i += 1;
             } else if (b < 0xE0) {
                 i += 2;
